Remove the machina finite state machine. Refactor imports in init.ts for improved clarity. Comment out unused script tags in index.html to clean up the code.
