+++ frontend/source/fsm/DiscographFSM.ts
@@ -1,0 +1,726 @@
+/**
+ * @fileoverview Implementation of the Discograph Finite State Machine
+ * using the state pattern and actions
+ */
+
+import * as d3 from "d3";
+import { discographManager, networkManager } from "../core";
+import {
+    restartForceLayout,
+    stopForceLayout,
+    displayForceLayout,
+    setupForceSliders,
+    startForceLayout,
+} from "../network/forceLayout";
+import type { NodeKey, NetworkCenter, NetworkData } from "../network/data";
+import {
+    processAPINetworkDataResponse,
+    convertNetworkDataToSimData,
+    updateGlobalData,
+} from "../network/data";
+import { pruneSimData } from "../network/pruning";
+import type { RelationsData } from "../relations";
+import { fetchAPINetwork, fetchAPIRandom, fetchAPIRadial } from "../api";
+import { resetNetworkTransform } from "../network/init";
+import { ALPHA } from "../network/forceLayout";
+import type { SimNode, SimLink } from "../network/data";
+import { RequestNetworkEvent, SelectEntityEvent } from "../network/events";
+import { showMessage } from "../messages";
+import { loading } from "../loading";
+import { FSM, INIT } from "../constants";
+import type { Actions } from "./actions/Actions";
+import type { State, StateContext } from "./State";
+import { ViewingNetworkState } from "./states/ViewingNetworkState";
+import { RequestingNetworkState } from "./states/RequestingNetworkState";
+import { RequestingRadialState } from "./states/RequestingRadialState";
+import { RequestingRandomState } from "./states/RequestingRandomState";
+import { ViewingRadialState } from "./states/ViewingRadialState";
+import { UninitializedState } from "./states/UninitializedState";
+import { debounce } from "../utils";
+import type { FSMStateType } from "./types";
+
+/**
+ * Type for event data that can be passed to handlers
+ */
+export type EventData =
+    | NetworkData
+    | RelationsData
+    | NetworkCenter
+    | NodeKey
+    | null;
+
+/**
+ * Type for the transition function passed to states
+ */
+export type TransitionFunction = (state: FSMStateType) => void;
+
+/**
+ * Implementation of the FSM using the state pattern
+ */
+export class DiscographFSM implements Actions {
+    /**
+     * The current state of the FSM
+     */
+    private _state: State;
+
+    /**
+     * Map of all available states
+     */
+    private _states: Map<FSMStateType, State>;
+
+    /**
+     * The current state type
+     */
+    private _currentStateType: FSMStateType;
+
+    /**
+     * Event handlers for FSM events
+     */
+    private _eventHandlers: Map<
+        string,
+        Set<(event: string, data: EventData) => void>
+    >;
+
+    /**
+     * Handler for showNetwork event
+     */
+    private _showNetworkHandler?: (event: Event) => void;
+
+    /**
+     * Create a new FSM instance
+     */
+    constructor() {
+        this._states = new Map();
+        this._eventHandlers = new Map();
+
+        // Initialize with the uninitialized state
+        this._currentStateType = "uninitialized";
+
+        // Register all states
+        this.registerState("uninitialized", new UninitializedState());
+        this.registerState("state-viewing-network", new ViewingNetworkState());
+        this.registerState(
+            "state-requesting-network",
+            new RequestingNetworkState(),
+        );
+        this.registerState(
+            "state-requesting-radial",
+            new RequestingRadialState(),
+        );
+        this.registerState(
+            "state-requesting-random",
+            new RequestingRandomState(),
+        );
+        this.registerState("state-viewing-radial", new ViewingRadialState());
+
+        this._state = this.getOrCreateState("uninitialized");
+
+        // Initialize the FSM
+        this.initialize();
+    }
+
+    /**
+     * Get the current state type
+     */
+    get state(): FSMStateType {
+        return this._currentStateType;
+    }
+
+    /**
+     * Register a state with the FSM
+     */
+    private registerState(stateType: FSMStateType, state: State): void {
+        this._states.set(stateType, state);
+    }
+
+    /**
+     * Get a state by type, creating a default one if it doesn't exist
+     */
+    private getOrCreateState(stateType: FSMStateType): State {
+        const state = this._states.get(stateType);
+        if (!state) {
+            console.warn(
+                `State ${stateType} not found, using uninitialized state as fallback`,
+            );
+            // Fall back to uninitialized state if requested state isn't registered
+            const fallbackState = this._states.get("uninitialized");
+            if (!fallbackState) {
+                throw new Error(
+                    "Uninitialized state not found, FSM is in an invalid state",
+                );
+            }
+            return fallbackState;
+        }
+        return state;
+    }
+
+    /**
+     * Initialize the FSM with event listeners
+     */
+    private initialize(): void {
+        // Event handlers
+        window.addEventListener(FSM.EVENTS.REQUEST_NETWORK, (event: Event) => {
+            if (event instanceof RequestNetworkEvent && event.detail) {
+                const { entityKey, pushHistory } = event.detail;
+                this.requestNetwork(entityKey, pushHistory);
+            }
+        });
+
+        window.addEventListener(FSM.EVENTS.REQUEST_RANDOM, () => {
+            this.requestRandom();
+        });
+
+        window.addEventListener(FSM.EVENTS.SELECT_ENTITY, (event: Event) => {
+            if (event instanceof SelectEntityEvent && event.detail) {
+                const { entityKey, fixed } = event.detail;
+                this.selectEntity(entityKey, fixed);
+            }
+        });
+
+        window.addEventListener(FSM.EVENTS.SHOW_NETWORK, () => {
+            this.handle("show-network", null, false, false);
+        });
+
+        window.addEventListener(FSM.EVENTS.SHOW_RADIAL, () => {
+            this.handle("show-radial", null, false, false);
+        });
+
+        // Handle browser history navigation
+        window.onpopstate = (event: PopStateEvent) => {
+            const state = event?.state as { key: string } | null;
+            if (state?.key) {
+                window.dispatchEvent(new RequestNetworkEvent(state.key, false));
+            }
+        };
+
+        // Handle window resize events with debounce
+        const handleResize = debounce(() => {
+            console.log("handleResize fsm");
+
+            // Center the loading visualization
+            const transform = `translate(${discographManager.svgDimensions[0] / 2},${discographManager.svgDimensions[1] / 2})`;
+            d3.selectAll(".centered")
+                .transition()
+                .duration(INIT.DEBOUNCE_DELAY)
+                .attr("transform", transform);
+
+            // Center the main node
+            if (networkManager.data.center) {
+                const centerNode = networkManager.data.nodeMap.get(
+                    networkManager.data.center.key,
+                );
+                if (centerNode) {
+                    centerNode.x = networkManager.newNodeCoords[0];
+                    centerNode.y = networkManager.newNodeCoords[1];
+                }
+            }
+
+            // Restart force layout if in network view
+            if (this.state === "state-viewing-network") {
+                console.log("restartForceLayout fsm");
+                restartForceLayout(ALPHA);
+            }
+        }, INIT.DEBOUNCE_DELAY);
+
+        window.addEventListener(FSM.EVENTS.RESIZE, handleResize);
+
+        // Handle SVG mousedown events
+        const svgDocument = document.getElementById("svg");
+        if (svgDocument) {
+            svgDocument.addEventListener("mousedown", () => {
+                if (this.state === "state-viewing-network") {
+                    this.selectEntity(null, false);
+                } else if (this.state === "state-viewing-radial") {
+                    this.handle("show-network", null, false, false);
+                }
+            });
+        }
+
+        // Initialize application state
+        this.loadInlineData();
+        this.toggleRadial(false);
+    }
+
+    /**
+     * Transition to a new state
+     */
+    transition(newStateType: FSMStateType): void {
+        console.log(
+            `Transitioning from ${this._currentStateType} to ${newStateType}`,
+        );
+
+        const context: StateContext = {
+            actions: this,
+            transition: this.transition.bind(this) as TransitionFunction,
+        };
+
+        // Exit the current state
+        this._state.onExit(context);
+
+        // Update state
+        this._currentStateType = newStateType;
+        this._state = this.getOrCreateState(newStateType);
+
+        // Enter the new state
+        this._state.onEnter(context);
+
+        // Emit state change event
+        this.emit("*", this._currentStateType);
+    }
+
+    /**
+     * Handle an event with optional data
+     */
+    handle(
+        event: string,
+        data: NetworkData | RelationsData | NetworkCenter | NodeKey | null,
+        pushHistory: boolean,
+        fixed: boolean,
+    ): void {
+        console.log(
+            `Handling event ${event} in state ${this._currentStateType}`,
+            data,
+        );
+
+        const context: StateContext = {
+            actions: this,
+            transition: this.transition.bind(this) as TransitionFunction,
+        };
+
+        // Map events to state methods
+        switch (event) {
+            case "received-network":
+                this._state.receivedNetwork?.(
+                    context,
+                    data as NetworkData,
+                    pushHistory,
+                );
+                break;
+            case "received-radial":
+                this._state.receivedRadial?.(context, data as RelationsData);
+                break;
+            case "received-random":
+                this._state.receivedRandom?.(context, data as NetworkCenter);
+                break;
+            case "request-network":
+                this._state.requestNetwork?.(context, data as NodeKey);
+                break;
+            case "request-random":
+                this._state.requestRandom?.(context);
+                break;
+            case "show-network":
+                this._state.showNetwork?.(context);
+                break;
+            case "show-radial":
+                this._state.showRadial?.(context);
+                break;
+            case "select-entity":
+                this._state.selectEntity?.(context, data as NodeKey, fixed);
+                break;
+            case "errored":
+                this._state.handleError?.(context, data);
+                break;
+            default:
+                console.warn(`Unhandled event: ${event}`);
+        }
+
+        // Emit the event
+        this.emit(event, data);
+    }
+
+    /**
+     * Register an event handler
+     */
+    on(event: string, handler: (event: string, data: EventData) => void): void {
+        if (!this._eventHandlers.has(event)) {
+            this._eventHandlers.set(event, new Set());
+        }
+
+        this._eventHandlers.get(event)?.add(handler);
+    }
+
+    /**
+     * Emit an event to registered handlers
+     */
+    private emit(event: string, data: EventData): void {
+        // Call handlers for the specific event
+        this._eventHandlers.get(event)?.forEach((handler) => {
+            handler(event, data);
+        });
+
+        // Call handlers for the wildcard event
+        if (event !== "*") {
+            this._eventHandlers.get("*")?.forEach((handler) => {
+                handler(event, data);
+            });
+        }
+    }
+
+    //
+    // Action implementations (from Actions interface)
+    //
+
+    /**
+     * Handle an error that occurred
+     */
+    handleError(error: unknown): void {
+        if (error instanceof Error) {
+            showMessage(error.message, "error");
+        } else {
+            showMessage("An unknown error occurred", "error");
+        }
+
+        this.transition("state-viewing-network");
+    }
+
+    /**
+     * Load data from the inline dgNetwork global variable
+     */
+    loadInlineData(): void {
+        if (window.dgNetwork) {
+            this.transition("state-requesting-network");
+
+            const networkData = processAPINetworkDataResponse(window.dgNetwork);
+
+            this.handle("received-network", networkData, false, false);
+        }
+    }
+
+    /**
+     * Update browser history state
+     */
+    pushState(entityKey: NodeKey, params?: Record<string, unknown>): void {
+        const [entityType, entityId] = entityKey.split("-");
+        const title = document.title;
+        let url = `/${entityType}/${entityId}`;
+
+        if (params) {
+            const searchParams = new URLSearchParams();
+            Object.entries(params).forEach(([key, value]) => {
+                if (Array.isArray(value)) {
+                    searchParams.set(key, value.join(","));
+                } else if (typeof value === "object" && value !== null) {
+                    // Convert objects to a stable JSON string
+                    const sorted = Object.fromEntries(
+                        Object.entries(value).sort(([a], [b]) =>
+                            a.localeCompare(b),
+                        ),
+                    );
+                    searchParams.set(key, JSON.stringify(sorted));
+                } else if (
+                    typeof value === "string" ||
+                    typeof value === "number" ||
+                    typeof value === "boolean"
+                ) {
+                    searchParams.set(key, String(value));
+                }
+            });
+            url += `?${decodeURIComponent(searchParams.toString())}`;
+        }
+
+        const state = { key: entityKey, params };
+        window.history.pushState(state, title, url);
+    }
+
+    /**
+     * Request network data for an entity
+     */
+    requestNetwork(entityKey: NodeKey, pushHistory: boolean): void {
+        this.transition("state-requesting-network");
+
+        fetchAPINetwork(entityKey)
+            .then((apiNetworkDataResponse) => {
+                const networkData = processAPINetworkDataResponse(
+                    apiNetworkDataResponse,
+                );
+
+                this.handle(
+                    "received-network",
+                    networkData,
+                    pushHistory,
+                    false,
+                );
+            })
+            .catch((error) => {
+                console.error("Error fetching network data:", error);
+
+                this.handleError(error);
+            });
+    }
+
+    /**
+     * Request radial data for an entity
+     */
+    requestRadial(entityKey: NodeKey): void {
+        this.transition("state-requesting-radial");
+
+        fetchAPIRadial(entityKey)
+            .then((relationsData) => {
+                this.handle("received-radial", relationsData, false, false);
+            })
+            .catch((error) => {
+                console.error("Error fetching radial data:", error);
+
+                this.handleError(error);
+            });
+    }
+
+    /**
+     * Request a random entity
+     */
+    requestRandom(): void {
+        this.transition("state-requesting-network");
+
+        fetchAPIRandom()
+            .then((networkCenter) => {
+                this.handle("received-random", networkCenter, true, false);
+            })
+            .catch((error) => {
+                console.error("Error fetching random data:", error);
+
+                this.handleError(error);
+            });
+    }
+
+    /**
+     * Display the network view
+     */
+    showNetwork(networkData: NetworkData, pushHistory: boolean): void {
+        console.log("showNetwork networkData: ", networkData);
+
+        this.transition("state-viewing-network");
+
+        if (!networkData || !networkData.center?.key) {
+            console.error("Invalid network data: missing center key");
+            return;
+        }
+
+        const filterSelect =
+            document.querySelector<HTMLSelectElement>("#filter select");
+        const filterValue = filterSelect?.value;
+        const params = { roles: Array.isArray(filterValue) ? filterValue : [] };
+
+        // Update the network data
+        document.title = "Discograph2: " + networkData.center.name;
+        document.body.setAttribute("id", networkData.center.key);
+
+        if (pushHistory) {
+            this.pushState(networkData.center.key, params);
+        }
+
+        console.log("received-network convertNetworkDataToSimData");
+        const simData = convertNetworkDataToSimData(networkData);
+
+        const prunedSimData = pruneSimData(simData);
+
+        updateGlobalData(prunedSimData);
+
+        console.log("received-network resetNetworkTransform");
+        resetNetworkTransform();
+
+        console.log("received-network startForceLayout");
+        startForceLayout(Array.from(prunedSimData.nodeMap.values()));
+
+        displayForceLayout();
+
+        setupForceSliders();
+
+        restartForceLayout(ALPHA);
+
+        this.handle("select-entity", networkData.center.key, false, false);
+    }
+
+    /**
+     * Display the radial view
+     */
+    showRadial(data?: RelationsData): void {
+        this.transition("state-viewing-radial");
+        this.handle("show-radial", data, false, false);
+    }
+
+    /**
+     * Toggle filter visibility
+     */
+    toggleFilter(show: boolean): void {
+        console.log("toggleFilter", show);
+
+        if (show === true) {
+            d3.select("#filter-container").classed("hidden", false);
+            const transform = `translate(${discographManager.svgDimensions[0] / 2},${discographManager.svgDimensions[1] / 2})`;
+            d3.select("#filter-container").attr("transform", transform);
+        } else {
+            d3.select("#filter-container").classed("hidden", true);
+        }
+    }
+
+    /**
+     * Toggle network visibility
+     */
+    toggleNetwork(status: boolean): void {
+        if (status) {
+            const root = networkManager.layers.root;
+            if (root) {
+                root.style("transition", "opacity 250ms").style("opacity", 1);
+            }
+        } else {
+            stopForceLayout();
+            const root = networkManager.layers.root;
+            if (root) {
+                root.style("transition", "opacity 250ms").style(
+                    "opacity",
+                    0.25,
+                );
+            }
+        }
+    }
+
+    /**
+     * Toggle loading indicator
+     */
+    toggleLoading(status: boolean): void {
+        loading.toggle(status);
+    }
+
+    /**
+     * Toggle radial view
+     */
+    toggleRadial(status: boolean): void {
+        const entityRelations = document.getElementById("entity-relations");
+        if (!entityRelations) {
+            console.warn("entity-relations element not found in DOM");
+            return;
+        }
+
+        if (status) {
+            if (this._showNetworkHandler) {
+                entityRelations.removeEventListener(
+                    "click",
+                    this._showNetworkHandler,
+                );
+            }
+
+            this._showNetworkHandler = (event: Event) => {
+                this.handle("show-network", null, false, false);
+                event.preventDefault();
+            };
+
+            entityRelations.addEventListener("click", this._showNetworkHandler);
+        } else {
+            if (this._showNetworkHandler) {
+                entityRelations.removeEventListener(
+                    "click",
+                    this._showNetworkHandler,
+                );
+            }
+
+            this._showNetworkHandler = (event: Event) => {
+                this.handle("show-radial", null, false, false);
+                event.preventDefault();
+            };
+
+            entityRelations.addEventListener("click", this._showNetworkHandler);
+        }
+    }
+
+    /**
+     * Select an entity in the network
+     */
+    selectEntity(entityKey: NodeKey | null, fixed: boolean): void {
+        console.log("selectEntity", entityKey, fixed);
+        discographManager.selectedNodeKey = entityKey;
+        let nodeOn: d3.Selection<SVGGElement, SimNode, SVGGElement, unknown>;
+        let nodeOff: d3.Selection<SVGGElement, SimNode, SVGGElement, unknown>;
+        let linkOn: d3.Selection<SVGGElement, SimLink, SVGGElement, unknown>;
+        let linkOff: d3.Selection<SVGGElement, SimLink, SVGGElement, unknown>;
+
+        const nodeLayer = networkManager.layers.node;
+        if (!nodeLayer) {
+            console.log("Network node layer not found");
+            return;
+        }
+
+        const linkLayer = networkManager.layers.link;
+        if (!linkLayer) {
+            console.log("Network link layer not found");
+            return;
+        }
+
+        if (entityKey !== null) {
+            nodeOn = nodeLayer.selectAll<SVGGElement, SimNode>(
+                "g" + "#node-" + entityKey,
+            );
+            nodeOff = nodeLayer.selectAll<SVGGElement, SimNode>(
+                "g.node:not(#node-" + entityKey + ")",
+            );
+
+            if (nodeOn.empty()) {
+                console.log("nodeOn not found");
+                return;
+            }
+            if (nodeOff.empty()) {
+                console.log("nodeOff not found");
+                return;
+            }
+
+            const nodeData = nodeOn.datum();
+            if (!nodeData) {
+                console.log("nodeData not found");
+                return;
+            }
+
+            console.log("nodeData: ", nodeData);
+            const linkKeys = nodeData.links.map((l) => l.key);
+            const linkSelection = linkLayer.selectAll<SVGGElement, SimLink>(
+                "g.link",
+            );
+
+            linkOn = linkSelection.filter((d: SimLink) =>
+                linkKeys.includes(d.key),
+            );
+            linkOff = linkSelection.filter(
+                (d: SimLink) => !linkKeys.includes(d.key),
+            );
+
+            const node = networkManager.data.nodeMap.get(entityKey);
+            if (!node) {
+                console.log("node not found");
+                return;
+            }
+            console.log("new selected node: ", node);
+
+            const [, id] = node.key.split("-");
+            const url = `http://discogs.com/${node.type}/${id}`;
+
+            d3.select("#entity-name").text(node.name);
+            d3.select("#entity-link").attr("href", url);
+
+            const entityDetails = d3.select("#entity-details");
+            entityDetails.classed("hidden", false);
+            entityDetails.style("display", "block");
+
+            d3.select("#navbar-title").text(node.name);
+
+            nodeOn.raise();
+            nodeOn.classed("selected", true);
+            console.log("nodeOn: ", nodeOn);
+
+            if (fixed) {
+                //nodeOn.each(function(d) { d.fixed = true; });
+                node.fixed = true;
+            }
+            linkOn.classed("selected", true);
+        } else {
+            nodeOff = nodeLayer.selectAll<SVGGElement, SimNode>("g.node");
+            linkOff = linkLayer.selectAll<SVGGElement, SimLink>("g.link");
+        }
+
+        if (nodeOff) {
+            console.log("nodeOff: ", nodeOff);
+            nodeOff.classed("selected", false).each((d) => (d.fixed = false));
+        }
+        if (linkOff) {
+            console.log("linkOff: ", linkOff);
+            linkOff.classed("selected", false);
+        }
+    }
+}
