+++ frontend/source/fsm/states/BaseState.ts
@@ -1,0 +1,99 @@
+/**
+ * @fileoverview Base state implementation that other states can extend
+ */
+
+import type { NodeKey, NetworkData } from "../../network/data";
+import type { RelationsData } from "../../relations";
+import type { State, StateContext } from "../State";
+
+/**
+ * Base state implementation with default no-op methods
+ * Concrete states can extend this class and override only the methods they need
+ */
+export abstract class BaseState implements State {
+    /**
+     * Called when entering this state
+     */
+    onEnter(_context: StateContext): void {
+        // Default implementation does nothing
+    }
+
+    /**
+     * Called when exiting this state
+     */
+    onExit(_context: StateContext): void {
+        // Default implementation does nothing
+    }
+
+    /**
+     * Handle an error that occurred in this state
+     */
+    handleError(context: StateContext, error: unknown): void {
+        // Default implementation delegates to actions
+        context.actions.handleError(error);
+    }
+
+    /**
+     * Handle a network data received event
+     */
+    receivedNetwork(
+        _context: StateContext,
+        _data: NetworkData,
+        _pushHistory = false,
+    ): void {
+        // Default implementation does nothing
+    }
+
+    /**
+     * Handle a radial data received event
+     */
+    receivedRadial(_context: StateContext, _data: RelationsData): void {
+        // Default implementation does nothing
+    }
+
+    /**
+     * Handle a random entity received event
+     */
+    receivedRandom(_context: StateContext, _data: { center: NodeKey }): void {
+        // Default implementation does nothing
+    }
+
+    /**
+     * Handle a request to show the network
+     */
+    showNetwork(_context: StateContext): void {
+        // Default implementation does nothing
+    }
+
+    /**
+     * Handle a request to show the radial view
+     */
+    showRadial(_context: StateContext): void {
+        // Default implementation does nothing
+    }
+
+    /**
+     * Handle a request to get a network for an entity
+     */
+    requestNetwork(_context: StateContext, _entityKey: NodeKey): void {
+        // Default implementation does nothing
+    }
+
+    /**
+     * Handle a request to get a random entity
+     */
+    requestRandom(_context: StateContext): void {
+        // Default implementation does nothing
+    }
+
+    /**
+     * Handle a request to select an entity
+     */
+    selectEntity(
+        _context: StateContext,
+        _entityKey: NodeKey | null,
+        _fixed: boolean,
+    ): void {
+        // Default implementation does nothing
+    }
+}
