+++ frontend/templates/index.html
@@ -15,9 +15,9 @@
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Open+Sans&display=swap">
 
     <!-- https://github.com/ifandelse/machina.js and dependencies -->
-    <script src="https://cdn.jsdelivr.net/npm/underscore@1.13.6/underscore-umd-min.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/lodash@4.17.21/lodash.min.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/machina@4.0.2/lib/machina.min.js"></script>
+    <!--    <script src="https://cdn.jsdelivr.net/npm/underscore@1.13.6/underscore-umd-min.js"></script>-->
+    <!--    <script src="https://cdn.jsdelivr.net/npm/lodash@4.17.21/lodash.min.js"></script>-->
+    <!--    <script src="https://cdn.jsdelivr.net/npm/machina@4.0.2/lib/machina.min.js"></script>-->
 
     <!-- jstree -->
     <link rel="stylesheet" href="//cdnjs.cloudflare.com/ajax/libs/jstree/3.3.8/themes/default/style.min.css"/>
